Remove the -u flag from ScalaTest testOptions (#55)
SA-51408

The "-u" flag, followed by a directory, instructed ScalaTest to report
test results in the specified directory.  However, newer versions of SBT
automatically enable a built-in JUnitXmlReportPlugin which write the
same reports.  These reporters can conflict with each other, with each
trying to append to the same file concurrently and producing invalid
data.

Either the built-in plugin could be disabled or the flag removed from
the ScalaTest testOptions setting in order to remove the conflict.  This
commit removes the ScalaTest version since that can be specified within
this plugin rather than in the destination project, and it represents
fewer deviations from the default settings overall.